[Feat] #28 - Consolidate component imports into a single module
