--- ahttp_client/header.py
@@ -1,47 +1,0 @@
-"""MIT License
-
-Copyright (c) 2023-present gunyu1019
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-"""
-
-from typing import Any
-from .component import Component
-
-
-class Header(Component):
-    """This class is used when a function's parameters are used as headers in an HTTP request.
-
-    Examples
-    --------
-    >>> def function(header: str | Header):
-    ...    pass
-    """
-
-    DEFAULT_KEY = "__DEFAULT_HEADER__"
-
-    @staticmethod
-    def default_header(key: str, value: Any):
-        def decorator(func):
-            if not hasattr(func, Header.DEFAULT_KEY):
-                setattr(func, Header.DEFAULT_KEY, dict())
-            getattr(func, Header.DEFAULT_KEY)[key] = value
-            return func
-
-        return decorator
